Bugs and Fixes
Bugs and fixes

Fixed incorrect map reading from ai (did not read the first 0 confirming
the start of the ai). TileX and TileY are required to be between 0 and
mapWidth-1 and the addtion of 1 will be made by the users and ai's if
needed. Game ends in frontend but the ship map needs to be cleaned up.